+++ src/components/app.tsx
@@ -3,11 +3,6 @@ import '../assets/stylesheets/app.scss'
 import React, { FC } from 'react'
 import DropzoneComponent from './dropzone-component'
 
-const App: FC = () => (
-  <>
-    <div className="app__title">Sandwich Beta</div>
-    <DropzoneComponent />
-  </>
-)
+const App: FC = () => (<DropzoneComponent />)
 
 export default App