+++ webpack.renderer.config.js
@@ -1,34 +1,8 @@
-const rules = require('./webpack.rules');
-const plugins = require('./webpack.plugins');
-
-rules.push({
-  test: /\.css$/,
-  use: ['style-loader', 'css-loader']
-});
-
-rules.push({
-  test: /\.s[ac]ss$/,
-  use: ['style-loader', 'css-loader', 'sass-loader']
-});
-
-rules.push({
-  test: /\.(woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
-  use: [
-    {
-      loader: 'file-loader',
-      options: {
-        name: '[name].[ext]',
-        outputPath: 'fonts/'
-      }
-    }
-  ]
-});
-
 module.exports = {
   module: {
-    rules
+    rules: require('./webpack.rules')
   },
-  plugins: plugins,
+  plugins: require('./webpack.plugins'),
   resolve: {
     extensions: ['.js', '.ts', '.jsx', '.tsx', '.css']
   }