+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# 0.2.3 (2020-09-12)
+
+## New Features
+
+- Improved design.
+  - Text: "Make PDF" should be larger.
+  - "Make PDF" button should be larger.
+  - Text: "Drag files here" should be larger.
+  - The bottom/footer area is too large.
+  - When there's no files loaded, "Make PDF" button should be deactivated or not shown.
+  - Add grabbing style.
+  - Style thumbnail and list view.
+  - On hover show delete button for each file.
+
+## Bug Fixes
+
+- Window should be able to grab and move.
+- Grabbing the image in the box to drag causes it to crash.
+- Reduce grab delay.
+- Make dropzone always present.
+- Add "Add Files" button.
+- Fix Font Awesome loading issue.
+- Upgrade Node packages
+- Add reload to menu options.
+
 # 0.2.2 (2020-08-16)
 
 ## New Features